basic calc reimplemented as abstract class
