+++ src/app/components/calculator/calculators/physical-hit-calculator.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { CalculatorInterface, Mode } from './calculator.interface';
+import { BasicCalculator, Mode } from './basic-calculator';
 import { PhysicalHitChanceFactors } from './factors/physical-hit-chance-factor.interface';
 import { PhysicalDamageFactors } from './factors/physical-damage-factors.interface';
 
 @Injectable()
-export class PhysicalHitCalculatorService implements CalculatorInterface<PhysicalHitChanceFactors, PhysicalDamageFactors> {
-  readonly chanceConfig = {
+export class PhysicalHitCalculatorService extends BasicCalculator<PhysicalHitChanceFactors, PhysicalDamageFactors> {
+  protected readonly config = {
     default: {
       hitCoefficient: 45,
       maxDice: 95,
@@ -19,7 +19,7 @@ export class PhysicalHitCalculatorService implements CalculatorInterface<Physica
   };
 
   getModeSuccessChance(factors: PhysicalHitChanceFactors, mode: Mode) {
-    const config = this.chanceConfig[mode];
+    const config = this.config[mode];
     const criticalHit = config.hitCoefficient +
       (4 * factors.enemyLevel + factors.distance * 4) -
       (factors.agility + (factors.playerLevel * 3));
@@ -27,21 +27,7 @@ export class PhysicalHitCalculatorService implements CalculatorInterface<Physica
   }
 
   getModeDamage(factors: PhysicalDamageFactors, mode: Mode) {
-    const config = this.chanceConfig[mode];
+    const config = this.config[mode];
     return (factors.weaponDamage + (factors.strength * factors.diceValue) / 60) * config.multiplier;
-  }
-
-  getSuccessChance(factors: PhysicalHitChanceFactors) { // ugly
-    return {
-      default: this.getModeSuccessChance(factors, 'default'),
-      critical: this.getModeSuccessChance(factors, 'critical')
-    };
-  }
-
-  getDamage(factors: PhysicalDamageFactors) { // ugly
-    return {
-      default: this.getModeDamage(factors, 'default'),
-      critical: this.getModeDamage(factors, 'critical')
-    };
   }
 }
