+++ src/app/components/calculator/calculators/spell-calculator.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { CalculatorInterface, Mode } from './calculator.interface';
+import { BasicCalculator, Mode } from './basic-calculator';
 import { SpellDamageFactors } from './factors/spell-damage-factors.interface';
 import { SpellCastChance } from './factors/spell-cast-chance.interface';
 
 @Injectable()
-export class SpellCalculatorService implements CalculatorInterface<SpellCastChance, SpellDamageFactors> {
-  readonly chanceConfig = {
+export class SpellCalculatorService extends BasicCalculator<SpellCastChance, SpellDamageFactors> {
+  protected readonly config = {
     default: {
       spellCoefficient: 20,
       multiplier: 1,
@@ -17,27 +17,13 @@ export class SpellCalculatorService implements CalculatorInterface<SpellCastChan
   };
 
   getModeSuccessChance(factors: SpellCastChance, mode: Mode) {
-    const config = this.chanceConfig[mode];
+    const config = this.config[mode];
     return config.spellCoefficient * factors.spellLevel - 2 * factors.wisdom;
   }
 
   getModeDamage(factors: SpellDamageFactors, mode: Mode) {
-    const config = this.chanceConfig[mode];
+    const config = this.config[mode];
     return factors.spellDamage + ((factors.spellDamage / 100) * (factors.intelligence / 3 +
       ((factors.diceValue * config.multiplier + (factors.intelligence * 2)) / 3)));
-  }
-
-  getSuccessChance(factors: SpellCastChance) { // ugly
-    return {
-      default: this.getModeSuccessChance(factors, 'default'),
-      critical: this.getModeSuccessChance(factors, 'critical')
-    };
-  }
-
-  getDamage(factors: SpellDamageFactors) { // ugly
-    return {
-      default: this.getModeDamage(factors, 'default'),
-      critical: this.getModeDamage(factors, 'critical')
-    };
   }
 }
